refactor: extract managed service runtime
